Migration for changing type of phone column in the UserInfoEntity was added.

+++ app/src/main/java/ua/romanik/roomtask/presentation/ui/fragment/user/create/CreateUserViewModel.kt
@@ -36,7 +36,7 @@ class CreateUserViewModel(
                     UserDomainModel(
                         name = name,
                         email = email,
-                        phone = phone,
+                        phone = phone.toLong(),
                         departmentId = selectedDepartment.id
                     )
                 )
@@ -60,7 +60,7 @@ class CreateUserViewModel(
                         id = userId,
                         name = name,
                         email = email,
-                        phone = phone,
+                        phone = phone.toLong(),
                         departmentId = selectedDepartment.id
                     )
                 )
